+++ src/helpers/configuration.ts
@@ -44,7 +44,7 @@ class Configuration {
 
   public update(
     section: string,
-    value: any,
+    value: unknown,
     configurationTarget?: ConfigurationTarget | boolean
   ): Thenable<void> {
     return this.configuration.update(section, value, configurationTarget);